Make ButtonClick Sound from SoundManager
